Module Shuttle_http.Body

module Stream : sig ... end

type t = private

| Empty

| Fixed of string

| Stream of Stream.t

val sexp_of_t : t -> Sexplib0.Sexp.t

val string : string -> t

string str creates a fixed length encoded body from a user provided string.

val empty : t

empty is a zero length body.

val of_pipe : [ `Chunked | `Fixed of int ] -> string Async.Pipe.Reader.t -> t

of_pipe is a convenience function that creates a streaming body from a user provided Async_kernel.Pipe.Reader.t. The pipe will be closed whenever the streaming body is closed, or EOF is reached.

val stream : Stream.t -> t

stream creates a streaming body from a user provided streaming module.

val to_stream : t -> Stream.t

to_stream converts a HTTP body to a stream.