Module Client.Oneshot

val call : ?interrupt:unit Async.Deferred.t -> ?connect_timeout:Core.Time.Span.t -> ?ssl:Ssl.t -> Address.t -> Request.t -> Response.t Async.Deferred.t

call Performs a one-shot http client call to the user provided connection target. If ssl options are provided the client will attempt to setup a SSL connection. If ssl options contain a hostname then the client will perform hostname verification to ensure the hostnames on the peer's ssl certificate matches the hostname provided by the caller. To disable this check or to customize how the ssl certificate is validated users can provide their own implementation of verify_certificate when creating the ssl options.