Module Shuttle_http.Meth

type t = [

| `GET

| `HEAD

| `POST

| `PUT

| `DELETE

| `CONNECT

| `OPTIONS

| `TRACE

| `PATCH

]

Request method is used to indicate the purpose of a HTTP request.

See https://datatracker.ietf.org/doc/html/rfc7231#section-4.3 for more details.

val sexp_of_t : t -> Sexplib0.Sexp.t

val t_of_sexp : Sexplib0.Sexp.t -> t

val __t_of_sexp__ : Sexplib0.Sexp.t -> t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

include Ppx_enumerate_lib.Enumerable.S with type t := t

val all : t list

include Ppx_quickcheck_runtime.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val of_string : string -> t Core.Or_error.t

val to_string : t -> string

val is_safe : t -> bool

is_safe t returns true if the semantics for a HTTP method are essentially read-only, and the client does not expect any state change on the server as a result of the request.

See https://datatracker.ietf.org/doc/html/rfc7231#section-4.2.1 for more details.

val is_idempotent : t -> bool

is_idempotent t returns true if multiple requests with a HTTP method are intended to have the same effect on the server as a single such request. This function returns true for PUT, DELETE and all safe methods.

See https://datatracker.ietf.org/doc/html/rfc7231#section-4.2.2 for more details.

val is_cacheable : t -> bool

is_cacheable t indicates that responses to requests with an HTTP method are allowed to be strored for future reuse. This function returns true for GET, HEAD and POST.

See https://datatracker.ietf.org/doc/html/rfc7231#section-4.2.3 for more details.

include Core.Comparable.S with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : (t, comparator_witness) Base__Comparator.comparator

val validate_lbound : min:t Core__.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core__.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core__.Maybe_bound.t -> max:t Core__.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : sig ... end

module Set : sig ... end