Module Shuttle_http.Parser

type error =

| Partial

| Fail of Core.Error.t

type chunk_kind =

| Start_chunk

| Continue_chunk of int

type chunk_parser_result =

| Chunk_complete of string

| Done

| Partial_chunk of string * int

val parse_request : ?pos:int -> ?len:int -> Core.Bigstring.t -> (Request.t * int, error) Core.result

Attempts to parse a buffer into a HTTP request. If successful, it returns the parsed request and an offset value that indicates the starting point of unconsumed content left in the buffer.

val parse_chunk_length : ?pos:int -> ?len:int -> Core.Bigstring.t -> (int * int, error) Core.result

val parse_chunk : ?pos:int -> ?len:int -> Core.Bigstring.t -> chunk_kind -> (chunk_parser_result * int, error) Core.result

module Private : sig ... end