Module Body.Stream

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val of_pipe : [ `Chunked | `Fixed of int ] -> string Async.Pipe.Reader.t -> t

of_pipe is a convenience function that creates a stream from a user provided Async_kernel.Pipe.Reader.t. The pipe will be closed whenever the streaming body is closed, or EOF is reached.

val close : t -> unit

val encoding : t -> [ `Chunked | `Fixed of int ]

val iter : t -> f:(string -> unit Async.Deferred.t) -> unit Async.Deferred.t

val drain : t -> unit Async.Deferred.t

val closed : t -> unit Async.Deferred.t

val read_started : t -> bool